refactor(web): move Configuration Chat Agent to utility agents page
The Configuration Chat Agent setting was buried at the bottom of the
agents page. Move it to the utility agents page where it sits more
naturally alongside other agent-selection controls.
