+++ apps/web/app/settings/utility-agents/page.tsx
@@ -1,5 +1,11 @@
+import { ConfigChatAgentSection } from "@/components/settings/config-chat-agent-section";
 import { UtilityAgentsSection } from "@/components/settings/utility-agents-section";
 
 export default function UtilityAgentsSettingsPage() {
-  return <UtilityAgentsSection />;
+  return (
+    <div className="space-y-8">
+      <UtilityAgentsSection />
+      <ConfigChatAgentSection />
+    </div>
+  );
 }
